+++ NebulaCompatibilityAssist/src/Hotfix/NebulaHotfix.cs
@@ -1,14 +1,5 @@
 ﻿using HarmonyLib;
 using System;
-using System.Collections.Generic;
-
-using UnityEngine;
-using NebulaModel.Packets.Players;
-using NebulaWorld.MonoBehaviours.Remote;
-using NebulaPatcher.Patches.Dynamic;
-using NebulaWorld;
-using NebulaModel;
-using NebulaNetwork;
 
 namespace NebulaCompatibilityAssist.Hotfix
 {
@@ -25,13 +16,12 @@ namespace NebulaCompatibilityAssist.Hotfix
             try
             {
                 System.Version nebulaVersion = pluginInfo.Metadata.Version;
-                /*
-                if (nebulaVersion.Major == 0 && nebulaVersion.Minor == 8 && nebulaVersion.Build == 13)
+                
+                if (nebulaVersion.Major == 0 && nebulaVersion.Minor == 8 && nebulaVersion.Build == 14)
                 {
-                    Patch0813(harmony);
-                    Log.Info("Nebula hotfix 0.8.13 - OK");                    
+                    harmony.PatchAll(typeof(Warper0814));
+                    Log.Info("Nebula hotfix 0.8.14 - OK");
                 }
-                */
                 ChatManager.Init(harmony);
                 harmony.PatchAll(typeof(Analysis.StacktraceParser));
                 Log.Info("Nebula extra features - OK");
@@ -44,176 +34,11 @@ namespace NebulaCompatibilityAssist.Hotfix
         }
 
         /*
-        private static void Patch0813(Harmony harmony)
+        private static void PatchPacketProcessor(Harmony harmony)
         {
-            Type classType;
-            classType = AccessTools.TypeByName("NebulaWorld.Multiplayer");
+            Type classType = AccessTools.TypeByName("NebulaWorld.Multiplayer");
             harmony.Patch(AccessTools.Method(classType, "HostGame"), new HarmonyMethod(typeof(NebulaNetworkPatch).GetMethod(nameof(NebulaNetworkPatch.BeforeMultiplayerGame))));
             harmony.Patch(AccessTools.Method(classType, "JoinGame"), new HarmonyMethod(typeof(NebulaNetworkPatch).GetMethod(nameof(NebulaNetworkPatch.BeforeMultiplayerGame))));
-
-            harmony.PatchAll(typeof(Warper0813));
-        }
-
-        private static class Warper0813
-        {
-            #region mecha animation
-            struct Snapshot
-            {
-                public EMovementState MovementState { get; set; }
-                public float HorzSpeed { get; set; }
-                public float VertSpeed { get; set; }
-                public float Turning { get; set; }
-                public float JumpWeight { get; set; }
-                public float JumpNormalizedTime { get; set; }
-                public byte IdleAnimIndex { get; set; }
-                public byte MiningAnimIndex { get; set; }
-                public float MiningWeight { get; set; }
-                public PlayerMovement.EFlags Flags { get; set; }
-            }
-
-            static readonly Dictionary<RemotePlayerAnimation, Snapshot[]> dict = new();
-
-
-            [HarmonyPrefix, HarmonyPatch(typeof(RemotePlayerAnimation), "UpdateState")]
-            static bool RemotePlayerAnimationUpdateState(RemotePlayerAnimation __instance, PlayerMovement packet)
-            {
-                if (!dict.TryGetValue(__instance, out var snapshotBuffer))
-                {
-                    snapshotBuffer = new Snapshot[3];
-                    dict[__instance] = snapshotBuffer;
-                }
-
-                for (int i = 0; i < snapshotBuffer.Length - 1; ++i)
-                {
-                    snapshotBuffer[i] = snapshotBuffer[i + 1];
-                }
-
-                snapshotBuffer[snapshotBuffer.Length - 1] = new Snapshot()
-                {
-                    MovementState = packet.MovementState,
-                    HorzSpeed = packet.HorzSpeed,
-                    VertSpeed = packet.VertSpeed,
-                    Turning = packet.Turning,
-                    JumpWeight = packet.JumpWeight,
-                    JumpNormalizedTime = packet.JumpNormalizedTime,
-                    IdleAnimIndex = packet.IdleAnimIndex,
-                    MiningAnimIndex = packet.MiningAnimIndex,
-                    MiningWeight = packet.MiningWeight,
-                    Flags = packet.Flags
-                };
-
-                // TODO: Fix effect timing
-                __instance.remotePlayerEffects.UpdateState(packet);
-
-                return false;
-            }
-
-            [HarmonyPrefix, HarmonyPatch(typeof(RemotePlayerAnimation), "Update")]
-            static bool RemotePlayerAnimationUpdate(RemotePlayerAnimation __instance)
-            {
-                if (!dict.TryGetValue(__instance, out var snapshotBuffer))
-                    return false;
-
-                var snapshot = snapshotBuffer[0];
-
-                __instance.PlayerAnimator.jumpWeight = snapshot.JumpWeight;
-                __instance.PlayerAnimator.jumpNormalizedTime = snapshot.JumpNormalizedTime;
-                __instance.PlayerAnimator.idleAnimIndex = snapshot.IdleAnimIndex;
-                __instance.PlayerAnimator.sailAnimIndex = 0;
-                __instance.PlayerAnimator.miningWeight = snapshot.MiningWeight;
-                __instance.PlayerAnimator.miningAnimIndex = snapshot.MiningAnimIndex;
-
-                __instance.PlayerAnimator.movementState = snapshot.MovementState;
-                __instance.PlayerAnimator.horzSpeed = snapshot.HorzSpeed;
-                __instance.PlayerAnimator.turning = snapshot.Turning;
-                PlanetData localPlanet = GameMain.galaxy.PlanetById(__instance.rootMovement.localPlanetId);
-                __instance.altitudeFactor = (localPlanet == null) ? 1f : Mathf.Clamp01((__instance.transform.position.magnitude - localPlanet.realRadius - 7f) * 0.15f);
-
-                float deltaTime = Time.deltaTime;
-                __instance.CalculateMovementStateWeights(__instance.PlayerAnimator, deltaTime);
-                __instance.CalculateDirectionWeights(__instance.PlayerAnimator);
-
-                __instance.PlayerAnimator.AnimateIdleState(deltaTime);
-                if (__instance.PlayerAnimator.idleAnimIndex == 0)
-                {
-                    for (int i = 1; i < __instance.PlayerAnimator.idles.Length; i++)
-                    {
-                        __instance.PlayerAnimator.idles[i].weight = 0;
-                        __instance.PlayerAnimator.idles[i].normalizedTime = 0f;
-                    }
-                }
-                __instance.PlayerAnimator.AnimateRunState(deltaTime);
-                __instance.PlayerAnimator.AnimateDriftState(deltaTime);
-                __instance.AnimateFlyState(__instance.PlayerAnimator);
-                __instance.AnimateSailState(__instance.PlayerAnimator);
-
-                __instance.PlayerAnimator.AnimateJumpState(deltaTime);
-                __instance.PlayerAnimator.AnimateSkills(deltaTime);
-                __instance.AnimateRenderers(__instance.PlayerAnimator);
-
-                //__instance.remotePlayerEffects.UpdateState(packet);
-
-                return false;
-            }
-            #endregion
-
-            #region save
-
-            [HarmonyPrefix]
-            [HarmonyPatch(typeof(GameSave_Patch), "SaveCurrentGame_Prefix")]
-            public static bool SaveCurrentGame_Prefix(string saveName, ref bool __result)
-            {
-                if (Multiplayer.IsActive && Multiplayer.Session.LocalPlayer.IsHost)
-                {
-                    // temp revert sand count back to original value before saving if we sync it (see SimulatedWorld.SetupInitialPlayerState() )
-                    if (Config.Options.SyncSoil)
-                    {
-                        int tmp = GameMain.mainPlayer.sandCount;
-                        GameMain.mainPlayer.sandCount = Multiplayer.Session.LocalPlayer.Data.Mecha.SandCount;
-                        Multiplayer.Session.LocalPlayer.Data.Mecha.SandCount = tmp;
-                    }
-                    SaveManager.SaveServerData(saveName);
-                }
-
-                // Only save if in single player or if you are the host
-                __result = !Multiplayer.IsActive || Multiplayer.Session.LocalPlayer.IsHost;
-                return false;
-            }
-
-            [HarmonyPrefix]
-            [HarmonyPatch(typeof(UIPerformancePanel), nameof(UIPerformancePanel.OnDataActiveButtonClick))]
-            public static bool OnDataActiveButtonClick_Prefix()
-            {
-                if (Multiplayer.IsActive)
-                {
-                    InGamePopup.ShowInfo("Access Denied", "Save test is not available in multiplayer game.", "OK");
-                    return false;
-                }
-                return true;
-            }
-
-            #endregion
-
-            [HarmonyPrefix]
-            [HarmonyPatch(typeof(PlanetData), nameof(PlanetData.LoadFactory))]
-            public static void LoadFactory_Prefix(PlanetData __instance)
-            {
-                if (!__instance.loading && __instance.factory != null)
-                {
-                    // Unload dummy physics & audio (remote planet timer < 10s) to fully initialize
-                    Log.Info("Clean " + __instance.displayName + "physics + audio");
-                    if (__instance.physics != null)
-                    {
-                        __instance.physics.Free();
-                        __instance.physics = null;
-                    }
-                    if (__instance.audio != null)
-                    {
-                        __instance.audio.Free();
-                        __instance.audio = null;
-                    }
-                }
-            }
         }
         */
     }