+++ NebulaCompatibilityAssist/src/Patches/DSP_Battle_Patch.cs
@@ -10,6 +10,7 @@ using System.Reflection;
 using System.Reflection.Emit;
 using System.Threading;
 using UnityEngine;
+using System.Collections.Concurrent;
 
 namespace NebulaCompatibilityAssist.Patches
 {
@@ -31,6 +32,7 @@ namespace NebulaCompatibilityAssist.Patches
                 NebulaModAPI.OnPlayerJoinedGame += Warper.OnPlayerJoinedGame;
                 NC_ModSaveData.OnReceive += Warper.ImportData;
                 harmony.PatchAll(typeof(Warper));
+                harmony.PatchAll(typeof(Warper_RemoteBuildings));
 
                 Log.Info($"{NAME} - OK");
                 NC_Patch.RequriedPlugins += " +" + NAME;
@@ -95,6 +97,7 @@ namespace NebulaCompatibilityAssist.Patches
                     else if (stage == 2)
                     {
                         EnemyShips.Export(w);
+                        Warper_RemoteBuildings.Export(w);
                     }
 
 
@@ -144,8 +147,8 @@ namespace NebulaCompatibilityAssist.Patches
                 }
                 else if (stage == 2)
                 {
-                    Log.Info("Import enemy ships");
                     EnemyShips.Import(r);
+                    Warper_RemoteBuildings.Import(r);
                 }
 
                 WaveStages.ResetCargoAccIncTable(DSP_Battle.Configs.extraSpeedEnabled && Rank.rank >= 5);
@@ -416,8 +419,8 @@ namespace NebulaCompatibilityAssist.Patches
 
                 // Client only put the station into distroyedStation to remove it from target
                 // The real building removing is performed by Host
-                Log.Debug("=========> Ship " + ship.shipIndex.ToString() + " landed at station " + ship.shipData.otherGId.ToString());
-                RemoveEntities.distroyedStation[ship.shipData.otherGId] = 0;
+                // Log.Debug("=========> Ship " + ship.shipIndex.ToString() + " landed at station " + ship.shipData.otherGId.ToString());
+                // RemoveEntities.distroyedStation[ship.shipData.otherGId] = 0;
                 ship.state = EnemyShip.State.distroyed;
                 ship.shipData.inc--;
                 if (ship.shipData.inc > 0)
@@ -436,7 +439,6 @@ namespace NebulaCompatibilityAssist.Patches
                     using var p = NebulaModAPI.GetBinaryWriter();
                     var w = p.BinaryWriter;
                     w.Write(ship.shipData.planetB);
-                    w.Write(station.entityId);
                     w.Write(ship.damageRange);
                     w.Write(station.id);
                     var packet = new NC_BattleEvent(NC_BattleEvent.EType.RemoveEntities, NebulaModAPI.MultiplayerSession.LocalPlayer.Id, p.CloseAndGetBytes());
@@ -450,7 +452,6 @@ namespace NebulaCompatibilityAssist.Patches
             public static void SyncRemoveEntities(BinaryReader r)
             {
                 int planetId = r.ReadInt32();
-                int entityId = r.ReadInt32();
                 int damageRange = r.ReadInt32();
                 int id = r.ReadInt32();
 
@@ -459,8 +460,12 @@ namespace NebulaCompatibilityAssist.Patches
                 PlanetFactory factory = planet.factory;
                 if (factory == null)
                     return;
-                Vector3 stationPos = factory.entityPool[entityId].pos;
+                StationComponent station = factory.transport.stationPool[id];
+                Vector3 stationPos = factory.entityPool[station.entityId].pos;
 
+                RemoveEntities.removingComponets = true;
+                RemoveEntities.RemoveStation(factory, station);
+                RemoveEntities.removingComponets = false;
                 if (!RemoveEntities.pendingDestroyedEntities.ContainsKey(planetId))
                     RemoveEntities.pendingDestroyedEntities.Add(planetId, new List<Tuple<Vector3, int>>());
                 RemoveEntities.pendingDestroyedEntities[planetId].Add(new Tuple<Vector3, int>(stationPos, damageRange));
@@ -666,6 +671,17 @@ namespace NebulaCompatibilityAssist.Patches
 
             #region Sync EnemyShip state
 
+            [HarmonyPrefix, HarmonyPatch(typeof(EnemyShip), nameof(EnemyShip.BeAttacked))]
+            static bool BeAttacked(EnemyShip __instance, int atk, DamageType dmgType, ref int __result)
+            {
+                if (NC_Patch.IsClient)
+                {
+                    __result = CalculateAttack(__instance, atk, dmgType);
+                   return false;
+                }
+                return true;
+            }
+
             [HarmonyPrefix, HarmonyPatch(typeof(EnemyShips), nameof(EnemyShips.RemoveShip))]
             static bool RemoveShip()
             {
@@ -681,10 +697,18 @@ namespace NebulaCompatibilityAssist.Patches
             [HarmonyPostfix, HarmonyPatch(typeof(EnemyShip), nameof(EnemyShip.FindAnotherStation))]
             static void OnTargetChange(EnemyShip __instance)
             {
-                if (NebulaModAPI.IsMultiplayerActive && NebulaModAPI.MultiplayerSession.LocalPlayer.IsHost)
+                if (NebulaModAPI.IsMultiplayerActive)
                 {
-                    Log.Dev($"[Battle] ship retarget [{__instance.shipIndex}]:{__instance.state} P{__instance.shipData.planetB} HP:{__instance.hp}");
-                    SendEnemyShipState(__instance);
+                    Log.Dev($"[Battle] ship retarget [{__instance.shipIndex}]:{__instance.state} P:{__instance.shipData.planetB} gid:{__instance.shipData.otherGId}");
+
+                    if (NebulaModAPI.MultiplayerSession.LocalPlayer.IsHost)
+                    {
+                        SendEnemyShipState(__instance);
+                    }
+                    else // 客戶端重新找目標(塔被拆?)
+                    {
+                        __instance.maxSpeed = 0f; // 定住飛船, 等待主機同步新的目標
+                    }
                 }
             }
 
@@ -714,7 +738,8 @@ namespace NebulaCompatibilityAssist.Patches
                 w.Write(ship.shipData.stage);
                 w.Write(ship.shipData.otherGId);
                 w.Write(ship.shipData.planetB);
-                
+                w.Write(ship.maxSpeed);
+
                 // Positions
                 w.Write(ship.shipData.direction);
                 w.Write((float)ship.shipData.uPos.x);
@@ -745,7 +770,7 @@ namespace NebulaCompatibilityAssist.Patches
                         ship.hp = r.ReadInt32();
                         ship.shipData.inc = r.ReadInt32();
 
-                        Log.Dev($"[Battle] ship[{ship.shipIndex}]:{ship.state} HP:{ship.hp} INC:{ship.shipData.inc}");
+                        //Log.Dev($"[Battle] ship[{ship.shipIndex}]:{ship.state} HP:{ship.hp} INC:{ship.shipData.inc}");
 
                         if (ship.state == EnemyShip.State.distroyed)
                         {
@@ -769,6 +794,7 @@ namespace NebulaCompatibilityAssist.Patches
                         ship.shipData.stage = r.ReadInt32();
                         ship.shipData.otherGId = r.ReadInt32();
                         ship.shipData.planetB = r.ReadInt32();
+                        ship.maxSpeed = r.ReadSingle();
 
                         Log.Dev($"[Battle] ship[{ship.shipIndex}]:{ship.state} HP:{ship.hp} INC:{ship.shipData.inc}");
 
@@ -782,6 +808,56 @@ namespace NebulaCompatibilityAssist.Patches
                 catch
                 {
                     Log.Warn($"[Battle] error when updating ship {shipIndex}:{state}");
+                }
+            }
+
+
+            static int CalculateAttack(EnemyShip __instance, int atk, DamageType dmgType) // [COPY] EnemyShip.BeAttacked 前面計算傷害的部分
+            {
+                try
+                {
+                    lock (__instance)
+                    {
+                        if (__instance.state != EnemyShip.State.active) return 0;
+
+                        double bonus = 0;
+                        if (DSP_Battle.Configs.isEnemyWeakenedByRelic) // relic1-3 战斗最开始的1min受到伤害增加
+                            bonus += 0.3;
+                        if (Relic.HaveRelic(2, 12) && Relic.Verify(0.1)) // relic2-12 有概率暴击
+                            bonus += 1;
+                        if (Relic.HaveRelic(0, 2) && Relic.relic0_2Version == 1)
+                            bonus += 0.0003 * Relic.relic0_2Charge;
+                        atk = Relic.BonusDamage(atk, bonus);
+                                                
+                        // 精英波次减伤
+                        if (DSP_Battle.Configs.nextWaveElite == 1)
+                        {
+                            int shipType = DSP_Battle.Configs.enemyIntensity2TypeMap[__instance.intensity];
+                            if (shipType == 1 && dmgType == DamageType.bullet && Utils.RandDouble() > 0.1)
+                                atk = 0;
+                            else if (shipType == 3 && (dmgType == DamageType.missileAoe || dmgType == DamageType.mega))
+                                atk = (int)(0.1 * atk);
+                            else if (shipType == 4 && (dmgType == DamageType.laser || dmgType == DamageType.mega || dmgType == DamageType.shield))
+                                atk = (int)(0.2 * atk);
+                        }
+                        if (Relic.GetCursedRelicCount() > 0) // 被诅咒的圣物的负面效果
+                        {
+                            atk = (int)(atk * (1 - Relic.GetCursedRelicCount() * 0.05));
+                        }
+                        if (atk >= __instance.hp) // [MOD] 即將被擊落時, 免疫這次攻擊。留下殘血的敵船等主機事件回收
+                        {
+                            atk = 0;
+                        }
+                        __instance.hp -= atk;
+                        RelicFunctionPatcher.ApplyBloodthirster(atk);
+                        return atk;
+                    }
+                }
+                catch (Exception e)
+                {
+                    Log.Dev($"CalculateAttack error! hp:{__instance.hp} atk:{atk}");
+                    Log.Dev(e);
+                    return 0;
                 }
             }
 
@@ -852,8 +928,8 @@ namespace NebulaCompatibilityAssist.Patches
                 }
                 catch (Exception e)
                 {
-                    Log.Warn("AfterShipDestoryed error!");
-                    Log.Warn(e);
+                    Log.Dev("AfterShipDestoryed error!");
+                    Log.Dev(e);
                 }
             }
 
@@ -1427,6 +1503,681 @@ namespace NebulaCompatibilityAssist.Patches
             }
 
             #endregion
+        }
+
+        public static class Warper_RemoteBuildings
+        {
+            // Remote buildings : shadow ejector/silo 遠端建築: 影子砲塔
+
+            public static void Export(BinaryWriter w)
+            {
+                int count = 0;
+                foreach (var planet in GameMain.galaxy.stars[DSP_Battle.Configs.nextWaveStarIndex].planets)
+                {
+                    if (planet.factory == null) continue;
+                    w.Write(planet.id);
+
+                    var entityPool = planet.factory.entityPool;
+                    var factorySystem = planet.factory.factorySystem;
+
+                    var ejectorPool = factorySystem.ejectorPool;
+                    for (int i = 1; i < factorySystem.ejectorCursor; i++)
+                    {
+                        ref var ejector = ref ejectorPool[i];
+                        if (ejector.id == 0) continue;
+
+                        int protoId = entityPool[ejector.entityId].protoId;
+                        if (protoId == 2311) continue; // 不收錄原版彈射器
+
+                        // 記錄下EjectorPatch會用到的參數
+                        w.Write(protoId);
+                        w.Write(ejector.entityId);
+                        w.Write(ejector.bulletId);
+                        w.Write(ejector.bulletCount);
+                        w.Write(ejector.bulletInc);
+                        w.Write(ejector.chargeSpend);
+                        w.Write(ejector.coldSpend);
+                        w.Write(ejector.localPosN.x);
+                        w.Write(ejector.localPosN.y);
+                        w.Write(ejector.localPosN.z);
+                        count++;
+                    }
+                    w.Write(-1);
+
+                    var siloPool = factorySystem.siloPool;
+                    for (int i = 1; i < factorySystem.siloCursor; i++)
+                    {
+                        ref var silo = ref siloPool[i];
+                        if (silo.id == 0) continue;
+
+                        int protoId = entityPool[silo.entityId].protoId;
+                        if (protoId == 2312 || protoId == 8036) continue; // 不收錄原版彈射器
+
+                        // 記錄下SiloPatch會用到的參數
+                        w.Write(protoId);
+                        w.Write(silo.bulletId);
+                        w.Write(silo.bulletCount);
+                        w.Write(silo.bulletInc);
+                        w.Write(silo.chargeSpend);
+                        w.Write(silo.coldSpend);
+                        w.Write(silo.localPos.x);
+                        w.Write(silo.localPos.y);
+                        w.Write(silo.localPos.z);
+                        count++;
+                    }
+                    w.Write(-1);
+                }
+                w.Write(-1);
+                Log.Info("RemoteBuildings: Export buildings " + count);
+            }
+
+            static int ejectorCursor = 0;
+            static EjectorComponent[] ejectorPool = new EjectorComponent[32];
+            static int siloCursor = 0;
+            static SiloComponent[] siloPool = new SiloComponent[32];
+
+            public static void Import(BinaryReader r)
+            {
+                ejectorCursor = 1;
+                siloCursor = 1;
+
+                while (true)
+                {
+                    int planetId = r.ReadInt32();
+                    if (planetId == -1) break;
+                    bool hasFactory = GameMain.galaxy.PlanetById(planetId)?.factory != null;
+
+                    while (true)
+                    {
+                        int protoId = r.ReadInt32();
+                        if (protoId == -1) break;
+
+                        if (ejectorCursor >= ejectorPool.Length)
+                        {
+                            EjectorComponent[] oldPool = ejectorPool;
+                            ejectorPool = new EjectorComponent[oldPool.Length * 2];
+                            Array.Copy(oldPool, siloPool, oldPool.Length);
+                        }
+
+                        ejectorPool[ejectorCursor].planetId = planetId;
+                        ejectorPool[ejectorCursor].id = protoId; // Use id to store protoId
+                        ejectorPool[ejectorCursor].entityId = r.ReadInt32(); // 分辨目標用
+                        ejectorPool[ejectorCursor].bulletId = r.ReadInt32();
+                        ejectorPool[ejectorCursor].bulletCount = r.ReadInt32();
+                        ejectorPool[ejectorCursor].bulletInc = r.ReadInt32();
+                        ejectorPool[ejectorCursor].chargeSpend = r.ReadInt32();
+                        ejectorPool[ejectorCursor].coldSpend = r.ReadInt32();
+                        ejectorPool[ejectorCursor].localPosN = new Vector3(r.ReadSingle(), r.ReadSingle(), r.ReadSingle());
+
+                        if (!hasFactory) // 只有在工廠尚未載入時才加入影子建築
+                        {
+                            ejectorPool[ejectorCursor].localRot = Maths.SphericalRotation(ejectorPool[ejectorCursor].localPosN, 0f); // 以localPos猜測localRot
+                            ++ejectorCursor;
+
+                        }
+                    }
+
+                    while (true)
+                    {
+                        int protoId = r.ReadInt32();
+                        if (protoId == -1) break;
+
+                        if (siloCursor >= siloPool.Length)
+                        {
+                            SiloComponent[] oldPool = siloPool;
+                            siloPool = new SiloComponent[oldPool.Length * 2];
+                            Array.Copy(oldPool, siloPool, oldPool.Length);
+                        }
+
+                        siloPool[siloCursor].planetId = planetId;
+                        siloPool[siloCursor].id = protoId; // Use id to store protoId
+                        siloPool[siloCursor].bulletId = r.ReadInt32();
+                        siloPool[siloCursor].bulletCount = r.ReadInt32();
+                        siloPool[siloCursor].bulletInc = r.ReadInt32();
+                        siloPool[siloCursor].chargeSpend = r.ReadInt32();
+                        siloPool[siloCursor].coldSpend = r.ReadInt32();
+                        siloPool[siloCursor].localPos = new Vector3(r.ReadSingle(), r.ReadSingle(), r.ReadSingle());
+                        //Log.Debug($"[{siloCursor}] {protoId} : planet{siloPool[siloCursor].planetId} count{siloPool[siloCursor].bulletCount}");
+
+                        if (!hasFactory)
+                        {
+                            siloPool[siloCursor].localRot = Maths.SphericalRotation(siloPool[siloCursor].localPos, 0f);
+                            ++siloCursor;
+                        }
+                    }
+                }
+                Log.Info($"RemoteBuildings: Import ejector {ejectorCursor - 1} silo {siloCursor - 1}");
+            }
+
+            [HarmonyPostfix, HarmonyPatch(typeof(WorkerThreadExecutor), nameof(WorkerThreadExecutor.AssemblerPartExecute))]
+            static void UpdateShadowBuildings(int ___usedThreadCnt, int ___curThreadIdx)
+            {
+                // 在客戶端更新影子建築, 製造子彈/火箭動畫。 假設: 1.電力全滿 2.子彈保持為全空/全滿
+                if (!NC_Patch.IsClient) return;
+                if (DSP_Battle.Configs.nextWaveState != 3) return; // 只在戰鬥階段時運行
+
+                try
+                {
+                    int start, end;
+                    int localplanetId = GameMain.localPlanet?.id ?? -1;
+                    var dysonSphere = GameMain.data.dysonSpheres[DSP_Battle.Configs.nextWaveStarIndex];
+                    var dysonSwarm = dysonSphere?.swarm;
+
+                    if (dysonSwarm != null && WorkerThreadExecutor.CalculateMissionIndex(1, ejectorCursor - 1, ___usedThreadCnt, ___curThreadIdx, 4, out start, out end))
+                    {
+                        var astroPoses = GameMain.galaxy.astrosData;
+                        for (int i = start; i < end; i++)
+                        {
+                            ref var ejector = ref ejectorPool[i];
+                            if (ejector.planetId == localplanetId) continue; // 如果玩家已經登陸星球, 則跳過
+
+                            int bulletCount = ejector.bulletCount;
+                            EjectorPatch(ref ejector, dysonSwarm, astroPoses);
+                            ejector.bulletCount = bulletCount; // 鎖定子彈數量
+                        }
+                    }
+
+                    if (dysonSphere != null && WorkerThreadExecutor.CalculateMissionIndex(1, siloCursor - 1, ___usedThreadCnt, ___curThreadIdx, 4, out start, out end))
+                    {
+                        for (int i = start; i < end; i++)
+                        {
+                            ref var silo = ref siloPool[i];
+                            if (silo.planetId == localplanetId) continue; // 如果玩家已經登陸星球, 則跳過
+
+                            int bulletCount = silo.bulletCount;
+                            SiloPatch(ref silo, dysonSphere);
+                            silo.bulletCount = bulletCount; // 鎖定子彈數量
+                        }
+                    }
+                }
+                catch (Exception e)
+                {
+                    Log.Dev(e);
+                }
+            }
+
+            public static bool SiloPatch(ref SiloComponent __instance, DysonSphere sphere)
+            {
+                float power = 1.0f; // 默認滿電狀態
+                int planetId = __instance.planetId;
+                int starIndex = planetId / 100 - 1;
+                //int gmProtoId = __instance.id; // 使用id儲存
+
+                if (DSP_Battle.Configs.developerMode)
+                {
+                    __instance.bulletId = 8006;
+                    __instance.bulletCount = 99;
+                }
+                if (GameMain.instance.timei % 60 == 0 && __instance.bulletCount == 0)
+                {
+                    __instance.bulletId = MissileSilo.nextBulletId(__instance.bulletId);
+                }
+
+                if (__instance.fired && __instance.direction != -1)
+                {
+                    __instance.fired = false;
+                }
+                float num = (float)Cargo.accTableMilli[__instance.incLevel];
+                int num2 = (int)(power * 10000f * (1f + num) + 0.1f);
+                Mutex dysonSphere_mx = sphere.dysonSphere_mx;
+                uint result;
+                lock (dysonSphere_mx)
+                {
+                    //下面设定目标，发射时是选择最近目标；如果目标丢失则再随机选择目标
+                    int targetIndex = 0;
+                    //if (DSP_Battle.Configs.nextWaveState == 3) // MOD: 鎖定在戰鬥階段執行
+                    {
+                        targetIndex = MissileSilo.FindTarget(starIndex, planetId);
+                    }
+
+                    __instance.hasNode = (sphere.GetAutoNodeCount() > 0);
+                    if (targetIndex <= 0)  //if (!__instance.hasNode) 原本是没有节点，因此不发射
+                    {
+                        __instance.autoIndex = 0;
+                        if (__instance.direction == 1)
+                        {
+                            __instance.time = (int)((long)__instance.time * (long)__instance.coldSpend / (long)__instance.chargeSpend);
+                            __instance.direction = -1;
+                        }
+                        if (__instance.direction == -1)
+                        {
+                            __instance.time -= num2;
+                            if (__instance.time <= 0)
+                            {
+                                __instance.time = 0;
+                                __instance.direction = 0;
+                            }
+                        }
+                        if (power >= 0.1f)
+                        {
+                            result = 1U;
+                        }
+                        else
+                        {
+                            result = 0U;
+                        }
+                    }
+                    else if (power < 0.1f)
+                    {
+                        if (__instance.direction == 1)
+                        {
+                            __instance.time = (int)((long)__instance.time * (long)__instance.coldSpend / (long)__instance.chargeSpend);
+                            __instance.direction = -1;
+                        }
+                        result = 0U;
+                    }
+                    else
+                    {
+                        uint num3 = 0U;
+                        bool flag2;
+                        num3 = ((flag2 = (__instance.bulletCount > 0)) ? 3U : 2U);
+                        if (__instance.direction == 1)
+                        {
+                            if (!flag2)
+                            {
+                                __instance.time = (int)((long)__instance.time * (long)__instance.coldSpend / (long)__instance.chargeSpend);
+                                __instance.direction = -1;
+                            }
+                        }
+                        else if (__instance.direction == 0 && flag2)
+                        {
+                            __instance.direction = 1;
+                        }
+                        if (__instance.direction == 1)
+                        {
+                            __instance.time += num2;
+                            if (__instance.time >= __instance.chargeSpend)
+                            {
+                                AstroData[] astroPoses = sphere.starData.galaxy.astrosData;
+                                __instance.fired = true;
+                                //DysonNode autoDysonNode = sphere.GetAutoDysonNode(__instance.autoIndex + __instance.id); //原本获取目标节点，现在已不需要
+                                DysonRocket dysonRocket = default(DysonRocket);
+                                dysonRocket.planetId = __instance.planetId;
+                                dysonRocket.uPos = astroPoses[__instance.planetId].uPos + Maths.QRotateLF(astroPoses[__instance.planetId].uRot, __instance.localPos + __instance.localPos.normalized * 6.1f);
+                                dysonRocket.uRot = astroPoses[__instance.planetId].uRot * __instance.localRot * Quaternion.Euler(-90f, 0f, 0f);
+                                dysonRocket.uVel = dysonRocket.uRot * Vector3.forward;
+                                dysonRocket.uSpeed = 0f;
+                                dysonRocket.launch = __instance.localPos.normalized;
+                                //sphere.AddDysonRocket(dysonRocket, autoDysonNode); //原本
+                                int rocketIndex = MissileSilo.AddDysonRockedGniMaerd(ref sphere, ref dysonRocket, null); //这是添加了一个目标戴森球节点为null的火箭，因此被判定为导弹
+
+                                MissileSilo.MissileTargets[starIndex][rocketIndex] = targetIndex;
+                                MissileSilo.missileProtoIds[starIndex][rocketIndex] = __instance.bulletId;
+                                int damage = 0;
+                                if (__instance.bulletId == 8004) damage = DSP_Battle.Configs.missile1Atk;
+                                else if (__instance.bulletId == 8005) damage = DSP_Battle.Configs.missile2Atk;
+                                else if (__instance.bulletId == 8006) damage = DSP_Battle.Configs.missile3Atk;
+                                //注册导弹
+                                UIBattleStatistics.RegisterShootOrLaunch(__instance.bulletId, damage);
+
+                                __instance.autoIndex++;
+                                if (!Relic.HaveRelic(1, 5))
+                                {
+                                    __instance.bulletInc -= __instance.bulletInc / __instance.bulletCount;
+                                    __instance.bulletCount--;
+                                }
+                                if (__instance.bulletCount == 0)
+                                {
+                                    __instance.bulletInc = 0;
+                                }
+                                __instance.time = __instance.coldSpend;
+                                __instance.direction = -1;
+                            }
+                        }
+                        else if (__instance.direction == -1)
+                        {
+                            __instance.time -= num2;
+                            if (__instance.time <= 0)
+                            {
+                                __instance.time = 0;
+                                __instance.direction = (flag2 ? 1 : 0);
+                            }
+                        }
+                        else
+                        {
+                            __instance.time = 0;
+                        }
+                        result = num3;
+                    }
+                }
+                return false;
+            }
+
+            public static bool EjectorPatch(ref EjectorComponent __instance, DysonSwarm swarm, AstroData[] astroPoses)
+            {
+                float power = 1.0f;
+                int gmProtoId = __instance.id;
+
+                //子弹需求循环
+                if (__instance.bulletCount == 0 && gmProtoId != 8014 && GameMain.instance.timei % 60 == 0)
+                {
+                    __instance.bulletId = Cannon.nextBulletId(__instance.bulletId);
+                }
+                else if (gmProtoId == 8014)
+                {
+                    __instance.bulletId = 8007;
+                    __instance.bulletCount = 0;
+                }
+
+                __instance.targetState = EjectorComponent.ETargetState.None;
+
+                //下面是因为 炮需要用orbitId记录索敌模式，而orbitId有可能超出已设定的轨道数，为了避免溢出，炮的orbitalId在参与计算时需要独立指定为1。
+                //后续所有的__instance.orbitId都被替换为此
+                if (__instance.orbitId <= 0 || __instance.orbitId > 4)
+                {
+                    __instance.orbitId = 1;
+                }
+
+                CannonFire(ref __instance, power, swarm, astroPoses, gmProtoId);
+                return false;
+            }
+
+            private static uint CannonFire(ref EjectorComponent __instance, float power, DysonSwarm swarm, AstroData[] astroPoses, int gmProtoId)
+            {
+                int planetId = __instance.planetId;
+                int entityId = __instance.entityId;
+                int starIndex = planetId / 100 - 1;
+                int calcOrbitId = __instance.orbitId;
+                uint result = 0;
+
+                float num2 = (float)Cargo.incTableMilli[__instance.incLevel];
+                int num3 = (int)(power * 10000f * (1f + num2) + 0.1f);
+
+                bool relic0_6Activated = Relic.HaveRelic(0, 6) && __instance.bulletId == 8001; // relic0-6 京级巨炮如果激活并且当前子弹确实是穿甲磁轨弹
+                if (relic0_6Activated) num3 = (int)(num3 * 0.1);
+
+                __instance.targetState = EjectorComponent.ETargetState.OK;
+                bool flag = true;
+                int num4 = __instance.planetId / 100 * 100;
+                float num5 = __instance.localAlt + __instance.pivotY + (__instance.muzzleY - __instance.pivotY) / Mathf.Max(0.1f, Mathf.Sqrt(1f - __instance.localDir.y * __instance.localDir.y));
+                Vector3 vector = new Vector3(__instance.localPosN.x * num5, __instance.localPosN.y * num5, __instance.localPosN.z * num5);
+                VectorLF3 vectorLF = astroPoses[__instance.planetId].uPos + Maths.QRotateLF(astroPoses[__instance.planetId].uRot, vector);
+                Quaternion q = astroPoses[__instance.planetId].uRot * __instance.localRot;
+                VectorLF3 uPos = astroPoses[num4].uPos;
+                VectorLF3 b = uPos - vectorLF;
+
+                List<EnemyShip> sortedShips = EnemyShips.sortedShips(calcOrbitId, starIndex, __instance.planetId);
+
+                //下面的参数根据是否是炮还是太阳帆的弹射器有不同的修改
+                double maxtDivisor = 5000.0; //决定子弹速度
+                int damage = 0;
+                int loopNum = sortedShips.Count;
+                double cannonSpeedScale = 1;
+                if (gmProtoId == 8012)
+                    cannonSpeedScale = 2;
+                EnemyShip curTarget = null;
+
+                if (__instance.bulletId == 8001)
+                {
+                    maxtDivisor = relic0_6Activated ? DSP_Battle.Configs.bullet4Speed : DSP_Battle.Configs.bullet1Speed * cannonSpeedScale; // relic0-6京级巨炮 还会大大加速此子弹速度
+                    damage = (int)DSP_Battle.Configs.bullet1Atk; //只有这个子弹能够因为引力弹射器而强化伤害。这个强化是不是取消了？
+                                                      //if (relic0_6Activated) // relic0-6 京级巨炮效果，由于这个伤害只在统计中计算为发射伤害，实际造成上海市还要再重新计算，因此这里不计算了，统计中记为发射了基础的伤害
+                                                      //    damage = Relic.BonusDamage(damage, 500); 
+                }
+                else if (__instance.bulletId == 8002)
+                {
+                    maxtDivisor = DSP_Battle.Configs.bullet2Speed * cannonSpeedScale;
+                    damage = DSP_Battle.Configs.bullet2Atk;
+                }
+                else if (__instance.bulletId == 8003)
+                {
+                    maxtDivisor = DSP_Battle.Configs.bullet3Speed * cannonSpeedScale;
+                    damage = DSP_Battle.Configs.bullet3Atk;
+                }
+                else if (__instance.bulletId == 8007)
+                {
+                    maxtDivisor = DSP_Battle.Configs.bullet4Speed; //没有速度加成
+                    damage = DSP_Battle.Configs.bullet4Atk;
+                }
+
+                //不该参与循环的部分，换到循环前了
+
+                bool flag2 = __instance.bulletCount > 0;
+                if (gmProtoId == 8014) //脉冲炮不需要子弹
+                    flag2 = true;
+                VectorLF3 vectorLF2 = VectorLF3.zero;
+
+                int begins = Cannon.indexBegins;
+                if (begins >= loopNum)
+                {
+                    Interlocked.Exchange(ref Cannon.indexBegins, 0);
+                    begins = 0;
+                }
+
+                bool needFindNewTarget = true;
+                EnemyShip lastTargetShip = null;
+                if (Cannon.cannonTargets.ContainsKey(planetId))
+                {
+                    if (Cannon.cannonTargets[planetId].ContainsKey(entityId))
+                    {
+                        int lastTargetShipIndex = Cannon.cannonTargets[planetId][entityId];
+                        if (EnemyShips.ships.ContainsKey(lastTargetShipIndex) && EnemyShips.ships[lastTargetShipIndex].state == EnemyShip.State.active)
+                        {
+                            lastTargetShip = EnemyShips.ships[lastTargetShipIndex];
+                            needFindNewTarget = false; // 老目标存在的话，在下面的循环中首先判断老目标是否合法（不被阻挡、俯仰角合适等）
+                        }
+                    }
+                }
+                for (int gm = begins; gm < loopNum && gm < begins + 3; gm++)
+                {
+                    if (!needFindNewTarget && gm > begins) // 说明上一个循环判定了原目标，且原目标无法作为合法目标，因此重新开始判定目标
+                    {
+                        needFindNewTarget = true;
+                        gm = begins;
+                    }
+
+                    //新增的，每次循环开始必须重置
+                    __instance.targetState = EjectorComponent.ETargetState.OK;
+                    flag = true;
+                    flag2 = __instance.bulletCount > 0;
+                    if (gmProtoId == 8014) //脉冲炮不需要子弹
+                        flag2 = true;
+                    else if (relic0_6Activated) // relic0-6 京级巨炮效果 每次消耗五发弹药
+                        flag2 = __instance.bulletCount > 4;
+
+                    int shipIdx = 0;//ship总表中的唯一标识：index
+                    EnemyShip targetShip = sortedShips[gm];
+                    if (!needFindNewTarget) // 如果原本的目标存在，则先判断原本的目标，此时在这个循环中，gm=begins的ship并没有真的被计算俯仰角等合法性判断，因此假若原本的目标失效，进入了下个循环后要根据情况重置gm=begins（见循环节开头）
+                    {
+                        targetShip = lastTargetShip;
+                    }
+                    vectorLF2 = targetShip.uPos;
+                    shipIdx = targetShip.shipIndex;
+
+                    if (needFindNewTarget && (!EnemyShips.ships.ContainsKey(shipIdx) || targetShip.state != EnemyShip.State.active)) continue;
+
+
+                    VectorLF3 vectorLF3 = vectorLF2 - vectorLF;
+                    __instance.targetDist = vectorLF3.magnitude;
+                    vectorLF3.x /= __instance.targetDist;
+                    vectorLF3.y /= __instance.targetDist;
+                    vectorLF3.z /= __instance.targetDist;
+                    Vector3 vector2 = Maths.QInvRotate(q, vectorLF3);
+                    __instance.localDir.x = __instance.localDir.x * 0.9f + vector2.x * 0.1f;
+                    __instance.localDir.y = __instance.localDir.y * 0.9f + vector2.y * 0.1f;
+                    __instance.localDir.z = __instance.localDir.z * 0.9f + vector2.z * 0.1f;
+                    if ((double)vector2.y < 0.08715574 || vector2.y > 0.8660254f)
+                    {
+                        __instance.targetState = EjectorComponent.ETargetState.AngleLimit;
+                        flag = false;
+                    }
+                    if (flag2 && flag)
+                    {
+                        for (int i = num4 + 1; i <= __instance.planetId + 2; i++)
+                        {
+                            if (i != __instance.planetId)
+                            {
+                                double num6 = (double)astroPoses[i].uRadius;
+                                if (num6 > 1.0)
+                                {
+                                    VectorLF3 vectorLF4 = astroPoses[i].uPos - vectorLF;
+                                    double num7 = vectorLF4.x * vectorLF4.x + vectorLF4.y * vectorLF4.y + vectorLF4.z * vectorLF4.z;
+                                    double num8 = vectorLF4.x * vectorLF3.x + vectorLF4.y * vectorLF3.y + vectorLF4.z * vectorLF3.z;
+                                    if (num8 > 0.0)
+                                    {
+                                        double num9 = num7 - num8 * num8;
+                                        num6 += 120.0;
+                                        if (num9 < num6 * num6)
+                                        {
+                                            flag = false;
+                                            __instance.targetState = EjectorComponent.ETargetState.Blocked;
+                                            break;
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                    if (EnemyShips.ships.ContainsKey(shipIdx) && EnemyShips.ships[shipIdx].state == EnemyShip.State.active && __instance.targetState != EjectorComponent.ETargetState.Blocked && __instance.targetState != EjectorComponent.ETargetState.AngleLimit)
+                    {
+                        curTarget = EnemyShips.ships[shipIdx]; //设定目标
+                        if (!Cannon.cannonTargets.ContainsKey(planetId))
+                        {
+                            Cannon.cannonTargets.TryAdd(planetId, new ConcurrentDictionary<int, int>());
+                            Cannon.cannonTargets[planetId].TryAdd(entityId, shipIdx);
+                        }
+                        else
+                        {
+                            Cannon.cannonTargets[planetId].AddOrUpdate(entityId, shipIdx, (x, y) => shipIdx);
+                        }
+                        if (EjectorUIPatch.needToRefreshTarget) //如果需要刷新目标
+                        {
+                            if (EjectorUIPatch.curEjectorPlanetId == __instance.planetId && EjectorUIPatch.curEjectorEntityId == __instance.entityId)
+                            {
+                                EjectorUIPatch.curTarget = curTarget;
+                            }
+                        }
+                        break;
+                    }
+                }
+                //如果没有船/船没血了，就不打炮了
+                if (curTarget == null)
+                {
+                    Interlocked.Add(ref Cannon.indexBegins, 3);
+                    flag = false; //本身是由于俯仰限制或路径被阻挡的判断，现在找不到目标而不打炮也算做里面
+                }
+                else if (curTarget != null && curTarget.hp <= 0)
+                {
+                    flag = false;
+                }
+                else if (curTarget.state != EnemyShip.State.active)
+                {
+                    flag = false;
+                }
+
+                bool flag3 = flag && flag2;
+                result = (flag2 ? (flag ? 4U : 3U) : 2U);
+                if (__instance.direction == 1)
+                {
+                    if (!flag3)
+                    {
+                        __instance.time = (int)((long)__instance.time * (long)__instance.coldSpend / (long)__instance.chargeSpend);
+                        __instance.direction = -1;
+                    }
+                }
+                else if (__instance.direction == 0 && flag3)
+                {
+                    __instance.direction = 1;
+                }
+
+
+                if (__instance.direction == 1)
+                {
+                    __instance.time += num3;
+                    if (__instance.time >= __instance.chargeSpend)
+                    {
+                        __instance.fired = true;
+                        VectorLF3 uBeginChange = vectorLF;
+                        int bulletCost = relic0_6Activated ? 5 : 1;
+
+                        int bulletIndex = -1;
+
+                        if (gmProtoId != 8014 || GameMain.instance.timei % 5 == 1) // 相位炮五帧才发一个，但是伤害x5
+                        {
+                            //下面是添加子弹
+                            bulletIndex = swarm.AddBullet(new SailBullet
+                            {
+                                maxt = (float)(__instance.targetDist / maxtDivisor),
+                                lBegin = vector,
+                                uEndVel = VectorLF3.Cross(vectorLF2 - uPos, swarm.orbits[calcOrbitId].up).normalized * Math.Sqrt((double)(swarm.dysonSphere.gravity / swarm.orbits[calcOrbitId].radius)), //至少影响着形成的太阳帆的初速度方向
+                                uBegin = uBeginChange,
+                                uEnd = vectorLF2
+                            }, calcOrbitId);
+                        }
+
+                        //设定子弹目标以及伤害，并注册伤害
+                        try
+                        {
+                            if (bulletIndex != -1)
+                                swarm.bulletPool[bulletIndex].state = 0; //设置成0，该子弹将不会生成太阳帆
+                        }
+                        catch (Exception)
+                        {
+                            DspBattlePlugin.logger.LogInfo("bullet info1 set error.");
+                        }
+                        UIBattleStatistics.RegisterShootOrLaunch(__instance.bulletId, damage, bulletCost);
+
+                        if (bulletIndex != -1)
+                            Cannon.bulletTargets[swarm.starData.index].AddOrUpdate(bulletIndex, curTarget.shipIndex, (x, y) => curTarget.shipIndex);
+
+                        //Main.logger.LogInfo("bullet info2 set error.");
+
+
+                        try
+                        {
+                            int bulletId = __instance.bulletId;
+                            if (bulletIndex != -1)
+                                Cannon.bulletIds[swarm.starData.index].AddOrUpdate(bulletIndex, bulletId, (x, y) => bulletId);
+                            // bulletIds[swarm.starData.index][bulletIndex] = 1;//后续可以根据子弹类型/炮类型设定不同数值
+                        }
+                        catch (Exception)
+                        {
+                            DspBattlePlugin.logger.LogInfo("bullet info3 set error.");
+                        }
+                        int bulletIncCost = 0;
+                        if (__instance.bulletCount != 0)
+                        {
+                            bulletIncCost = bulletCost * __instance.bulletInc / __instance.bulletCount;
+                            __instance.bulletInc -= bulletIncCost;
+                        }
+                        __instance.bulletCount -= bulletCost;
+                        if (gmProtoId == 8012 && Relic.HaveRelic(2, 3) && Relic.Verify(0.75)) // relic2-3 回声 概率回填弹药
+                        {
+                            __instance.bulletCount += 1;
+                            __instance.bulletInc += bulletIncCost / bulletCost;
+                        }
+                        if (__instance.bulletCount <= 0)
+                        {
+                            __instance.bulletInc = 0;
+                            __instance.bulletCount = 0;
+                        }
+                        __instance.time = __instance.coldSpend;
+                        __instance.direction = -1;
+
+                        //if (gmProtoId == 8014) //激光炮为了视觉效果，取消冷却阶段每帧都发射（不能简单地将charge和cold的spend设置为0，因为会出现除以0的错误）
+                        //    __instance.direction = 1;
+
+                    }
+                }
+
+                else if (__instance.direction == -1)
+                {
+                    __instance.time -= num3;
+                    if (__instance.time <= 0)
+                    {
+                        __instance.time = 0;
+                        __instance.direction = (flag3 ? 1 : 0);
+                    }
+                }
+                else
+                {
+                    __instance.time = 0;
+
+                }
+
+                return result;
+            }
+
         }
     }
 }